Add initial Gradle build framework targeting Java 21
This borrows the rest of the missing Gradle build components from the old WRIMS Repo, as well as making a few small adjustments to bump to
Java 21. This should get the project far enough that it can be imported into IDEs and build a jar properly. Notably, this does not yet implement GitHub Actions, that will happen in future commits.