Add initial Gradle build framework targeting Java 21
This borrows the rest of the missing Gradle build components from the
old WRIMS Repo, as well as making a few small adjustments to bump to
Java 21. This should get the project far enough that it can be imported
into IDEs properly, although it still won't build due to missing
dependencies for some 3rd party jars.