Adopt SLF4J Logging Framework, with Log4J Implementation (#159)
* Add `Slf4J`, with `Log4J` implementation

Add the dependencies to gradle, and configure log4j with a single console
logger/appender setup. This configuration can be expanded on in the future
to log to a JSON format with other loggers. Slf4J is also compatible with
adding observability later via OpenTelemetry.

* Use Slf4J to log read configuration settings.

Add a few methods to support prettier logging of the read configuration
values.

* Use Async logging appender, add dependency

* Add file based logging for better debugging

Separates logs into three categories:

- Root logger: catches >=DEBUG, saves to file in User directory
- Run File logger: catches >=INFO, saves to file in the Model directory
- Console logger: catches >=INFO, shows on the console.