Update `gurobi` dependency (#182)
* Update to `com.gurobi:gurobi:11.0.3`

Change repository to Maven Central for Gurobi, and change import to the correct package.

* Bugfix: tests failed without write access to root drive

The default value for project.dir of "" in the `ConfigUtilsTest` would
cause tests to fail is the user could not create the file `\XXX.log`.
The logging configuration file is also updated to create a specific
`.wrims/logs/runs` folder to track these logs that fall-through.